+++ MJ_FB_Frontend/src/api/client.ts
@@ -1,0 +1,71 @@
+// src/api/client.ts
+// Shared API client utilities
+import type { Role, UserRole } from '../types';
+
+// Read API base URL from environment or fall back to localhost
+export const API_BASE =
+  import.meta.env.VITE_API_BASE || 'http://localhost:4000';
+
+function getCsrfToken() {
+  return document.cookie
+    .split('; ')
+    .find(row => row.startsWith('csrfToken='))?.split('=')[1];
+}
+
+export async function apiFetch(
+  input: RequestInfo | URL,
+  init: RequestInit = {},
+) {
+  const headers = new Headers(init.headers || {});
+  const csrf = getCsrfToken();
+  if (csrf) headers.set('X-CSRF-Token', csrf);
+  init.headers = headers;
+
+  let res = await fetch(input, { credentials: 'include', ...init });
+  if (res.status === 401) {
+    try {
+      const data = await res.clone().json();
+      if (data?.message === 'Token expired') {
+        const refreshRes = await fetch(`${API_BASE}/auth/refresh`, {
+          method: 'POST',
+          credentials: 'include',
+        });
+        if (refreshRes.ok) {
+          res = await fetch(input, { credentials: 'include', ...init });
+        }
+      }
+    } catch {
+      // ignore
+    }
+  }
+  return res;
+}
+
+export interface LoginResponse {
+  role: Role;
+  name: string;
+  bookingsThisMonth?: number;
+  userRole?: UserRole;
+}
+
+export async function handleResponse(res: Response) {
+  if (!res.ok) {
+    let message = res.statusText;
+    let data: any = null;
+    try {
+      data = await res.json();
+      message = data.message || data.error || JSON.stringify(data);
+    } catch {
+      message = await res.text();
+    }
+    const err: any = new Error(message);
+    if (data) err.details = data;
+    throw err;
+  }
+  if (res.status === 204 || res.headers.get('Content-Length') === '0') {
+    return undefined;
+  }
+  const text = await res.text();
+  return text ? JSON.parse(text) : undefined;
+}
+
