+++ MJ_FB_Frontend/src/components/VolunteerBookingHistory.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { getMyVolunteerBookings } from '../api/api';
+import { getMyVolunteerBookings } from '../api/volunteers';
 import type { VolunteerBooking } from '../types';
 import { formatTime } from '../utils/time';
 import Page from './Page';
