+++ MJ_FB_Frontend/src/pages/UserDashboard.tsx
@@ -19,7 +19,7 @@ import {
 } from '@mui/material';
 import { EventAvailable, Announcement, History } from '@mui/icons-material';
 import FeedbackSnackbar from '../components/FeedbackSnackbar';
-import { getBookingHistory, getSlots, getHolidays, cancelBooking } from '../api/api';
+import { getBookingHistory, getSlots, getHolidays, cancelBooking } from '../api/bookings';
 import type { Slot, Holiday } from '../types';
 import { formatTime } from '../utils/time';
